+++ examples/ts/test.ts
@@ -12,9 +12,9 @@ const game = new DuckEngine.Game({
 	background: '#fff',
 });
 
-const main = new MainScene();
+const main = new MainScene(game);
 
-game.scenes.add(main);
+game.scenes.add([main]);
 
 game.start();